+++ lib/dashboard/mod_dependencies_widget.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:trios/chipper/utils.dart';
 import 'package:trios/mod_manager/version_checker.dart';
+import 'package:trios/models/enabled_mods.dart';
 import 'package:trios/models/mod_variant.dart';
 
 import '../mod_manager/mod_manager_logic.dart';
@@ -26,9 +27,10 @@ class ModDependenciesWidget extends ConsumerStatefulWidget {
 class _ModDependenciesWidgetState extends ConsumerState<ModDependenciesWidget> {
   @override
   Widget build(BuildContext context) {
-    final enabledMods = ref.watch(AppState.enabledModsFile).valueOrNull;
     final modVariants = ref.watch(AppState.modVariants).valueOrNull;
+    final mods = ref.watch(AppState.mods);
     final gameVersion = ref.watch(AppState.starsectorVersion).valueOrNull;
+    final enabledMods = ref.watch(AppState.enabledModsFile).valueOrNull?.filterOutMissingMods(mods);
     if (modVariants == null || enabledMods == null) return const SizedBox();
 
     final modVariant = widget.modVariant;
@@ -64,12 +66,12 @@ class _ModDependenciesWidgetState extends ConsumerState<ModDependenciesWidget> {
             children: [
               Text("Original game version",
                   style: theme.textTheme.labelMedium
-                      ?.copyWith(color: vanillaWarningColor.withOpacity(0.8))),
+                      ?.copyWith(color: ThemeManager.vanillaWarningColor.withOpacity(0.8))),
               Padding(
                 padding: const EdgeInsets.only(left: 8.0),
                 child: Text(modInfo.originalGameVersion ?? "",
                     style: theme.textTheme.labelMedium
-                        ?.copyWith(color: vanillaWarningColor)),
+                        ?.copyWith(color: ThemeManager.vanillaWarningColor)),
               ),
             ],
           ),
@@ -111,11 +113,11 @@ class _ModDependenciesWidgetState extends ConsumerState<ModDependenciesWidget> {
                   style: theme.textTheme.labelMedium?.copyWith(
                       color: switch (dependencyState) {
                     Satisfied _ => null,
-                    Missing _ => vanillaErrorColor,
+                    Missing _ => ThemeManager.vanillaErrorColor,
                     Disabled _ =>
-                      vanillaWarningColor, // Disabled means it's present, so we can just enable it.
-                    VersionInvalid _ => vanillaErrorColor,
-                    VersionWarning _ => vanillaWarningColor,
+                      ThemeManager.vanillaWarningColor, // Disabled means it's present, so we can just enable it.
+                    VersionInvalid _ => ThemeManager.vanillaErrorColor,
+                    VersionWarning _ => ThemeManager.vanillaWarningColor,
                   })),
             );
           }),
@@ -125,7 +127,7 @@ class _ModDependenciesWidgetState extends ConsumerState<ModDependenciesWidget> {
           Text(
               "Warning: this mod requires a different version of a mod that you have installed, but might run with this one.",
               style: theme.textTheme.labelMedium
-                  ?.copyWith(color: vanillaErrorColor)),
+                  ?.copyWith(color: ThemeManager.vanillaErrorColor)),
       ],
     );
   }