+++ lib/models/enabled_mods.dart
@@ -1,4 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:trios/models/mod.dart';
 
 part '../generated/models/enabled_mods.freezed.dart';
 part '../generated/models/enabled_mods.g.dart';
@@ -7,5 +8,14 @@ part '../generated/models/enabled_mods.g.dart';
 class EnabledMods with _$EnabledMods {
   const factory EnabledMods(Set<String> enabledMods) = _EnabledMods;
 
-  factory EnabledMods.fromJson(Map<String, dynamic> json) => _$EnabledModsFromJson(json);
+  factory EnabledMods.fromJson(Map<String, dynamic> json) =>
+      _$EnabledModsFromJson(json);
+}
+
+extension EnabledModsExtensions on EnabledMods {
+  EnabledMods filterOutMissingMods(List<Mod> mods) {
+    return EnabledMods(enabledMods
+        .where((enabledModId) => mods.any((mod) => mod.id == enabledModId))
+        .toSet());
+  }
 }